feat: 로그아웃 기능 구현 완료, 산기반 추천 내 서치바 에러 해결

+++ front/src/Common/Navbar/Navbar.tsx
@@ -33,8 +33,13 @@ function Navbar() {
 
   const logOUt = async () => {
     const res = await userApi.logOut(accessToken, refreshToken);
-    console.log(res.data);
-    sessionStorage.clear();
+    if (res.data === "Success") {
+      sessionStorage.clear();
+      navigate("/");
+      setToggleOn(!toggleOn);
+    } else {
+      console.log(res.data);
+    }
   };
 
   const [toggleOn, setToggleOn] = useState<boolean>(false);
